requirements.txt
nvidia-cublas-cu12==12.1.3.1
py-cpuinfo==9.0.0
Jinja2==3.1.5
click==8.1.8
setuptools==75.1.0
certifi==2024.12.14
wandb==0.19.1
peft==0.13.0
aiosignal==1.3.2
sentry-sdk==2.19.2
setproctitle==1.3.4
sentencepiece==0.2.0
protobuf==3.20.2
pyarrow==18.1.0
mdurl==0.1.2
accelerate==0.34.2
annotated-types==0.7.0
tensorboard==2.18.0
transformers==4.46.3
frozenlist==1.5.0
networkx==3.4.2
nvidia-cuda-nvrtc-cu12==12.1.105
mpmath==1.3.0
ninja==1.11.1.3
pandas==2.2.3
trl==0.11.1
bitsandbytes==0.45.0
psutil==6.1.1
tqdm==4.67.1
hf_transfer==0.1.8
triton==2.3.1
smmap==5.0.2
torch==2.3.1
tensorboard-data-server==0.7.2
pip==24.2
nvidia-curand-cu12==10.3.2.106
multiprocess==0.70.16
absl-py==2.1.0
tokenizers==0.20.0
evaluate==0.4.0
nvidia-cudnn-cu12==8.9.2.26
filelock==3.16.1
responses==0.18.0
markdown-it-py==3.0.0
nvidia-cusolver-cu12==11.4.5.107
deepspeed==0.15.1
rich==13.9.4
GitPython==3.1.44
Pygments==2.18.0
aiohttp==3.11.11
tzdata==2024.2
regex==2024.11.6
gitdb==4.0.12
shtab==1.7.1
attrs==24.3.0
Levenshtein==0.26.1
typing_extensions==4.12.2
urllib3==2.3.0
nvidia-ml-py==12.560.30
xxhash==3.5.0
hjson==3.1.0
numpy==2.2.1
six==1.17.0
msgpack==1.1.0
einops==0.8.0
tyro==0.9.5
Werkzeug==3.1.3
typeguard==4.4.1
python-Levenshtein==0.26.1
platformdirs==4.3.6
dill==0.3.8
huggingface-hub==0.27.0
nvidia-nvtx-cu12==12.1.105
yarl==1.18.3
RapidFuzz==3.11.0
charset-normalizer==3.4.1
scipy==1.14.1
MarkupSafe==3.0.2
async-timeout==5.0.1
grpcio==1.68.1
Markdown==3.7
packaging==24.2
docstring_parser==0.16
docker-pycreds==0.4.0
sympy==1.13.1
multidict==6.1.0
nvidia-cuda-cupti-cu12==12.1.105
propcache==0.2.1
aiohappyeyeballs==2.4.4
python-dateutil==2.9.0.post0
idna==3.10
datasets==3.2.0
wheel==0.44.0
safetensors==0.4.5
nvidia-nvjitlink-cu12==12.4.127
PyYAML==6.0.2
nvidia-cufft-cu12==11.0.2.54
fsspec==2024.9.0
pydantic_core==2.27.2
requests==2.32.3
pydantic==2.10.4
pytz==2024.2
nvidia-nccl-cu12==2.20.5
nvidia-cusparse-cu12==12.1.0.106
nvidia-cuda-runtime-cu12==12.1.105
backports.tarfile==1.2.0
importlib_resources==6.4.0
autocommand==2.2.2
jaraco.text==3.12.1
more-itertools==10.3.0
typing_extensions==4.12.2
inflect==7.3.1
tomli==2.0.1
jaraco.collections==5.1.0
zipp==3.19.2
jaraco.functools==4.0.1
importlib_metadata==8.0.0
jaraco.context==5.3.0